+++ python/test/vmafrc_test.py
@@ -865,7 +865,7 @@ class VmafrcQualityRunnerTest(unittest.TestCase):
             None, fifo_mode=False,
             delete_workdir=True,
             result_store=None,
-            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_v0.6.1neg.pkl")}
+            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_float_v0.6.1neg.pkl")}
         )
         self.runner.run(parallelize=False)
 
@@ -893,7 +893,7 @@ class VmafrcQualityRunnerTest(unittest.TestCase):
             None, fifo_mode=False,
             delete_workdir=True,
             result_store=None,
-            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_v0.6.1neg.pkl"),
+            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_float_v0.6.1neg.pkl"),
                            'adm_enhn_gain_limit': 1.2}
         )
         self.runner.run(parallelize=False)
@@ -920,11 +920,11 @@ class VmafrcQualityRunnerTest(unittest.TestCase):
             optional_dict={
                 'models': [
                     'path={}:name=vmaf'.format(VmafConfig.model_path("vmaf_float_v0.6.1.pkl")),
-                    'path={}:name=vmafneg'.format(VmafConfig.model_path("vmaf_v0.6.1neg.pkl")),
+                    'path={}:name=vmafneg'.format(VmafConfig.model_path("vmaf_float_v0.6.1neg.pkl")),
                 ]
             }
         )
-        with self.assertRaises(AssertionError, msg="vmaf_float_v0.6.1.pkl and vmaf_v0.6.1neg.pkl require the same fex with "
+        with self.assertRaises(AssertionError, msg="vmaf_float_v0.6.1.pkl and vmaf_float_v0.6.1neg.pkl require the same fex with "
                                                    "different input arguments, but the exception is not raised."):
             self.runner.run(parallelize=False)
 
@@ -939,12 +939,12 @@ class VmafrcQualityRunnerTest(unittest.TestCase):
             result_store=None,
             optional_dict={
                 'models': [
-                    'path={}:name=vmafneg'.format(VmafConfig.model_path("vmaf_v0.6.1neg.pkl")),
+                    'path={}:name=vmafneg'.format(VmafConfig.model_path("vmaf_float_v0.6.1neg.pkl")),
                     'path={}:name=vmaf'.format(VmafConfig.model_path("vmaf_float_v0.6.1.pkl")),
                 ]
             }
         )
-        with self.assertRaises(AssertionError, msg="vmaf_v0.6.1neg.pkl and vmaf_float_v0.6.1.pkl require different input "
+        with self.assertRaises(AssertionError, msg="vmaf_float_v0.6.1neg.pkl and vmaf_float_v0.6.1.pkl require different input "
                                                    "arguments for the same fex, but the exception is not raised."):
             self.runner.run(parallelize=False)
 
@@ -963,10 +963,10 @@ class VmafrcQualityRunnerTest(unittest.TestCase):
             delete_workdir=True,
             result_store=None,
             optional_dict={'disable_clip_score': True,
-                           'model_filepath': VmafConfig.test_resource_path("vmaf_v0.6.1neg.inconsist.pkl")}
+                           'model_filepath': VmafConfig.test_resource_path("vmaf_float_v0.6.1neg.inconsist.pkl")}
         )
         with self.assertRaises(AssertionError,
-                               msg="vmaf_v0.6.1neg.inconsist.pkl has inconsistent vif_enhn_gain_limit "
+                               msg="vmaf_float_v0.6.1neg.inconsist.pkl has inconsistent vif_enhn_gain_limit "
                                    "values for the same fex but the exception is not raised."):
             self.runner.run(parallelize=False)
 
@@ -984,7 +984,7 @@ class VmafrcQualityRunnerTest(unittest.TestCase):
             None, fifo_mode=False,
             delete_workdir=True,
             result_store=None,
-            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_v0.6.1neg.pkl"),
+            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_float_v0.6.1neg.pkl"),
                            'adm_enhn_gain_limit': 0.9}
         )
         with self.assertRaises(AssertionError,
@@ -1005,7 +1005,7 @@ class VmafrcQualityRunnerTest(unittest.TestCase):
             None, fifo_mode=False,
             delete_workdir=True,
             result_store=None,
-            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_v0.6.1neg.pkl"),
+            optional_dict={'disable_clip_score': True, 'model_filepath': VmafConfig.model_path("vmaf_float_v0.6.1neg.pkl"),
                            'vif_enhn_gain_limit': 0.9}
         )
         with self.assertRaises(AssertionError,