+++ python/vmaf/script/convert_vmaf_model_to_vmaf_no_enhn_gain_model.py
@@ -40,7 +40,7 @@ else:
     convert_vmaf_model_to_vmaf_no_enhn_gain_model(
         # VmafConfig.model_path('vmaf_float_v0.6.1.pkl'),
         './model/vmaf_float_v0.6.1.pkl',
-        # VmafConfig.workspace_path('model', 'vmaf_v0.6.1neg.pkl')
-        './model/vmaf_v0.6.1neg.pkl',
+        # VmafConfig.workspace_path('model', 'vmaf_float_v0.6.1neg.pkl')
+        './model/vmaf_float_v0.6.1neg.pkl',
     )
     print('Done.')